added casts of pointer difference to integer

+++ mach/proto/ncg/codegen.c
@@ -590,7 +590,7 @@ normalfailed:	if (stackpad!=tokpatlen) {
 		do {
 			npos=exactmatch=0;
 			for(rpp=reglist[propno];rp= *rpp; rpp++)
-				if (getrefcount(rp-machregs, FALSE)==0) {
+				if (getrefcount((int)(rp-machregs), FALSE)==0) {
 					pos[npos++] = rp-machregs;
 					if (eqtoken(&rp->r_contents,&token))
 						pos2[exactmatch++] = rp-machregs;
